Moved components to seperate page from map
