+++ App/frontend/src/routes/Base.vue
@@ -1,52 +1,16 @@
 <template>
 <div>
-  <sidebar-menu class="sidebar" :menu="menu" :hide-toggle="true" @item-click="onItemClick"/>
-  <div class="main-box">
-    <grid-layout
-      :layout.sync="panes"
-      :col-num="12"
-      :row-height="30"
-      :is-draggable="true"
-      :is-resizable="false"
-      :is-mirrored="false"
-      :vertical-compact="true"
-      :prevent-collision="false"
-      :margin="[10, 10]"
-      :use-css-transforms="true"
-    >
-      <grid-item v-for="(item,index) in panes"
-                 style="background-color: #27293D; border: 1px black; margin: 5px; color:white;"
-                 :x="item.x"
-                 :y="item.y"
-                 :w="item.w"
-                 :h="item.h"
-                 :i="item.i"
-                 :key="index">
-        <i class="fa fa-times exit-symbol" aria-hidden="true" @click="removePane(index)"></i>
-        {{item.i}}
-        <MapComponent v-if="item.componentType == 'Map'"></MapComponent>
-        <TestComponent v-if="item.componentType == 'speed'"></TestComponent>
-        <br><br>
-      </grid-item>
-    </grid-layout>
-  </div>
+  <sidebar-menu class="sidebar" ref="sidebar" :menu="menu" :hide-toggle="true" @item-click="onItemClick"/>
 </div>
 </template>
 
 <script>
-import axios from 'axios'
-import VueGridLayout from 'vue-grid-layout';
-import TestComponent from "@/components/TestComponent";
 
 export default {
   name: 'Base',
   components: {
-      GridLayout: VueGridLayout.GridLayout,
-      GridItem: VueGridLayout.GridItem,
-      TestComponent
   },
   mounted() {
-
   },
   data () {
     return {
@@ -59,12 +23,13 @@ export default {
         {
           title: 'Map',
           icon: 'fas fa-map',
-          href: '/Map',
+          href: '/map',
           componentType: 'None'
         },
         {
           title: 'Components',
           icon: 'fas fa-dice-d6',
+          href: '/info',
           child: [
             {
               title: 'Data',
@@ -81,76 +46,7 @@ export default {
           ],
           componentType: 'None'
         }
-      ],
-      //Gridlayout elements
-      numOfPanes: 0, //Tracks current number of panes
-      paneID: 0, //Pane ID always increases to avoid having multiple panes with same ID
-      panes: []
-    }
-  },
-  methods: {
-    onItemClick (event, item) {
-      console.log(item);
-      if (item.componentSize == "2BlockV") {
-        this.panes.push({
-          "x": 0,
-          "y": 0,
-          "w": 4,
-          "h": 12,
-          "i": String(this.paneID),
-          "componentType": String(item.componentType)
-        });
-        this.numOfPanes++;
-        this.paneID++;
-      }
-      else if (item.componentSize == "2BlockH") {
-        this.panes.push({
-          "x": 0,
-          "y": 0,
-          "w": 8,
-          "h": 7,
-          "i": String(this.paneID),
-          "componentType": String(item.componentType)
-        });
-        this.numOfPanes++;
-        this.paneID++;
-      }
-      else if(item.componentSize == "4Block"){
-        this.panes.push({
-          "x": 0,
-          "y": 0,
-          "w": 8,
-          "h": 12,
-          "i": String(this.paneID),
-          "componentType": String(item.componentType)
-        });
-        this.numOfPanes++;
-        this.paneID++;
-      }
-      else if(item.componentType != "None"){
-        this.panes.push({
-          "x": 0,
-          "y": 0,
-          "w": 4,
-          "h": 7,
-          "i": String(this.paneID),
-          "componentType": String(item.componentType)
-        });
-        this.numOfPanes++;
-        this.paneID++;
-      }
-    },
-    //Remove panes
-    removePane(index){
-      this.numOfPanes--;
-      this.panes.splice(index,1); //Remove pane and shift proceceding values to prevent holes in array
-    },
-    //Test call to backend
-    pingBackend: function() {
-      axios.get('http://127.0.0.1:5000/Ping')
-        .then((response) => {
-          this.msg = response.data
-        })
+      ]
     }
   }
 }
@@ -184,7 +80,6 @@ a {
   width: 300px;
   background-color: #41B682;
 }
-
 
 .exit-symbol{
   position: absolute;
